requirements.txt
SpeechRecognition 
googletrans==4.0.0-rc1 
pyaudio 
pyttsx3
gTTS
playsound
flask==2.0.1
flask-socketio==5.1.1
speechrecognition==3.8.1
googletrans==4.0.0-rc1
gtts==2.2.3
playsound==1.2.2
python-engineio==4.3.0
python-socketio==5.5.0
eventlet==0.33.0
